Patch from Gábor Ábel: suggestmgr - missing check (just like in ngram) caused a possible negative index into an array
SVNReadProperties
- temporary std::string usage fixed

DeadCodes
- several dead codes removed

fallthrough
- intentional fallthrough should be marked

XMessageBox
- MaxButtonStringSize wasn't big enough, so in CXDialogTemplate::LoadUserDefinedButtonStrings all _tcscpy risked buffer overrun (since udcs.sz* has a size of MAX_PATH).

BufferedOutFile
- InternalClose should set 'file' to INVALID_HANDLE_VALUE, because IsOpen decides based on it. Before the patch the destructor of CCacheFileOutBuffer could double call CloseHandle on the same handle. First it explicitly calls RemoveFile() calling in turn InternalClose(), then returns. When it returns the destructor of BufferedOutFile runs, calling InternalClose() once again which did not recognize that the file is not open anymore.)

LogDlg
- EditLogMessage had an int argument and passed it to a function expecting a size_t. I changed the signature of EditLogMessage to have a size_t argument.
- EditLogMessage was possibly called with -1 as argument, now it has a check against this. 

git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@22801 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDialog/LogDlg.h
@@ -161,7 +161,7 @@ private:
     BOOL IsDiffPossible (const CLogChangedPath& changedpath, svn_revnum_t rev);
     BOOL Open(bool bOpenWith, CString changedpath, svn_revnum_t rev);
     void EditAuthor(const std::vector<PLOGENTRYDATA>& logs);
-    void EditLogMessage(int index);
+    void EditLogMessage(size_t index);
     void DoSizeV1(int delta);
     void DoSizeV2(int delta);
     void AdjustMinSize();
